다시보기가 실제 판과 다르게 흘러가던 문제 — dt 를 처음부터 float32 로 굴린다
판은 float64 dt 로 도는데 다시보기는 dt 를 float32 로 저장한다. 그 차이가
상대오차 1e-7 쯤인데, 전기선 시뮬은 그보다 훨씬 예민하다: dt 를 1e-8 만
흔들어도 죽는 시점이 60.43초에서 78.63초로 바뀐다. 스폰 타이밍이 한 프레임
밀리면 그 뒤 패턴이 통째로 갈리기 때문이다.

실제 사례(Guest24414088, 78.63초)에서 다시보기가 60.43초에 죽었고, 저장된
입력을 그대로 돌려 60.43초를 그대로 재현했다. 기록 자체는 진짜다 —
dt 합계가 78.632초로 기록과 맞고 마지막 프레임까지 조작이 살아 있다.

Math.fround 로 애초에 저장할 값 그대로 굴리면 저장 왕복이 값을 안 깎아,
기록한 판과 다시보기가 한 치도 안 어긋난다.

+++ public/js/main.js
@@ -2309,8 +2309,15 @@ const pollen = scene.getObjectByName('pollen');
 function frame() {
   requestAnimationFrame(frame);
 
-  // 탭이 백그라운드에 갔다 오면 dt가 튀므로 상한을 둔다
-  const dt = Math.min(clock.getDelta(), 1 / 20);
+  // 탭이 백그라운드에 갔다 오면 dt가 튀므로 상한을 둔다.
+  //
+  // Math.fround 로 float32 로 깎아서 쓴다. 다시보기는 dt 를 float32 로 저장하는데
+  // (encodeReplay), 판이 돌 때는 float64 로 굴리면 저장된 값과 미세하게 어긋난다.
+  // 그 차이가 1e-7 쯤인데, 전기선 시뮬은 그보다 훨씬 예민하다 — dt 를 1e-8 만
+  // 흔들어도 죽는 시점이 60초에서 78초로 바뀌는 걸 확인했다(스폰 타이밍이 한
+  // 프레임 밀리면 그 뒤 패턴이 통째로 갈린다). 그래서 애초에 저장할 값 그대로
+  // 굴려, 기록한 판과 다시보기가 한 치도 안 어긋나게 한다.
+  const dt = Math.fround(Math.min(clock.getDelta(), 1 / 20));
   const now = performance.now() / 1000;
 
   animateAmbient(dt, now);
